Move get_session_data to step chooser
